+++ GUI/GUI.lua
@@ -16,14 +16,13 @@ local CreateAchievementPointsTooltip;
 function gui:LoadWithBlizzard_AchievementUI()
     self.SetAchievementFrameHeight(addon.Options.db.Window.AchievementFrameHeightOffset); -- Do this in order to create the correct amount of buttons based on our settings
 
-    gui.Calendar.Load();
-
     gui.AchievementsFrame:Load();
     addon.GUI.SummaryFrame:Load();
     gui.CategoriesFrame:Load();
     gui.FilterButton:Load();
 
     gui.Search:Load();
+    addon.GUI.Calendar:Load();
 
     gui.AddDataToBlizzardTabs();