+++ CHANGELOG.md
@@ -6,13 +6,19 @@ All notable changes to this project will be documented in this file.
 - Summary view for each new tab that shows the last 25 (default) earned achievements and the category's progress with optional ElvUI skin
 - Option to add your current locale to the Wowhead link (active by default)
 - Option to add a related tab to the Wowhead link
+- Days on the calendar with more than 4 achievements earned now show an arrow to have a visual indication for the user
+- Side window in the calendar with a list of all the earned achievements on the selected day
 
 ### Changed
 - Power Unbound and Power Unleashed are obtainable again
+- Calendar is now movable
 
 ### Fixed
 - A visual bug that would paint accountwide unobtainable achievements blue instead of red
 - When earning an achievement, the GUI should now properly reflect that
+
+### Removed
+- Achievements on the calendar view do not show their tooltip anymore and can't be clicked
 
 ## 9.2.5.35.2 - 2022-06-28
 ### Added